requirements.txt
python-telegram-bot==20.6
openpyxl==3.1.2
pandas==2.1.3
